Adjusted answers to Pointcut expressions question
