+++ src/main/java/spring/certification/aop/helper/HelperConfigurationMarker.java
@@ -1,0 +1,9 @@
+package spring.certification.aop.helper;
+
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+
+@ComponentScan
+@Configuration
+public class HelperConfigurationMarker {
+}
